add a new logging class, but don't use it yet
